backend/requirements.txt
# ===== Core LiveKit Agent Framework =====
livekit>=0.17.0
livekit-agents>=1.2.0

# ===== LiveKit Plugins =====
livekit-plugins-silero>=0.1.5
livekit-plugins-turn-detector>=0.1.3
livekit-plugins-openai>=0.5.0

# ===== Speech-to-Text (STT) =====
faster-whisper>=1.0.3
ctranslate2>=4.0.0  # GPU support for faster-whisper

# ===== Text-to-Speech (TTS) =====
edge-tts>=6.1.0
pydub>=0.25.1

# ===== Language Model (LLM) =====
ollama>=0.1.7

# ===== Audio Processing =====
soundfile>=0.12.1
numpy>=1.26.4

# ===== FastAPI Backend =====
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6

# ===== Database (PostgreSQL) =====
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
alembic>=1.13.0

# ===== HTTP Clients =====
requests>=2.31.0
httpx>=0.25.0

# ===== Utilities =====
python-dotenv>=1.0.1
python-dateutil>=2.8.2
orjson>=3.9.10
loguru>=0.7.2
typing-extensions>=4.9.0

# ===== Authentication =====
passlib[bcrypt]>=1.7.4
PyJWT>=2.8.0
bcrypt>=4.1.0
email-validator>=2.0.0

# ===== CORS Support =====
# (included in fastapi, no separate package needed)

# ===== Frontend =====
# Note: Frontend uses LiveKit JS SDK via CDN
# - livekit-client (JS): https://unpkg.com/livekit-client
# - Pure HTML/CSS/JavaScript - no Python dependencies
